+++ floatcsep/__init__.py
@@ -1,9 +1,8 @@
-from floatcsep import logger
-from floatcsep import registry
-from floatcsep import accessors
 from floatcsep import evaluation
 from floatcsep import experiment
 from floatcsep import model
-from floatcsep import readers
+from floatcsep.infrastructure import engine, environments, registries, repositories, logger
+from floatcsep.utils import readers, accessors, helpers
+from floatcsep.postprocess import reporting, plot_handler
 
 __version__ = "0.1.4"
