+++ setup.cfg
@@ -18,7 +18,10 @@ url = https://github.com/cseptesting/floatcsep.git
 [options]
 packages =
     floatcsep
-    floatcsep.cmd
+    floatcsep.commands
+    floatcsep.infrastructure
+    floatcsep.postprocess
+    floatcsep.utils
 install_requires =
     numpy
     dateparser
@@ -82,7 +85,7 @@ dev =
 
 [options.entry_points]
 console_scripts =
-    floatcsep = floatcsep.cmd.main:floatcsep
+    floatcsep = floatcsep.commands.main:floatcsep
 
 [flake8]
 max-line-length = 96
